simplify hash table code

+++ curunner/main.c
@@ -75,7 +75,7 @@ int main(int argc, const char* argv[]) {
     Summary result = { {0,0}, 0 };
     static char buffer[256];
     ACU_Files* files;
-    int allocs;
+    long allocs;
 
     ACU_FilesVisitor testExecuteVisitor = { execute , NULL };
     testExecuteVisitor.context = (void*) &result;
@@ -105,6 +105,6 @@ int main(int argc, const char* argv[]) {
     acu_enabledTrackMemory(0);
 
     allocs = acu_getAllocCount();
-    printf("\n\rallocs = %d", allocs);
+    printf("\n\rallocs = %ld", allocs);
     return result.returnValue == ACU_TEST_PASSED ? 0 : 2;
 }
